feat: link zip code search to forecast
The ZipCodeSearch component's search text is linked to the ForecastList. When a valid zipcode is input to the zip code search, a fetch to the Open Weather API is made to get the current weather data. When the fetch is resolved, the current display is updated to show the new locale's current weather situation.