feat: hourly forecast fetches
Tri-hourly forecasts are fetched by default instead of the current weather. Plans are to have an hourly forecast tab and a current weather tab with a shared search bar. Some refactoring on the current design will be necessary. Current issues in this commit: 1) times aren't formatted correctly in hourly forecast tabs. 2) flat list not allowing to scroll past item 6.